# docs/swagger.yaml
basePath: /v1
definitions:
  controllers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.VisitDto:
    properties:
      clientId:
        type: integer
      dateTime:
        description: Date and time.
        format: date-time
        type: string
      doctorId:
        type: integer
      info:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
      petId:
        type: integer
      serviceId:
        type: integer
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Client:
    properties:
      birthDate:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      info:
        type: string
      name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      sex:
        type: string
      surname:
        type: string
      updated_at:
        type: string
    type: object
  models.Department:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      slug:
        type: string
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.Pet:
    properties:
      breed:
        type: string
      client:
        $ref: '#/definitions/models.Client'
      clientId:
        type: integer
      colour:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      sex:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Service:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: integer
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      departments:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      active:
        type: boolean
      birthDate:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      departments:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      email:
        type: string
      id:
        type: integer
      info:
        type: string
      name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      profession:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      roleId:
        type: integer
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      sex:
        type: string
      slug:
        type: string
      surname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Visit:
    properties:
      client:
        $ref: '#/definitions/models.Client'
      clientId:
        type: integer
      created_at:
        type: string
      dateTime:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      info:
        type: string
      lastUpdatedBy:
        $ref: '#/definitions/models.User'
      lastUpdatedById:
        type: integer
      pet:
        $ref: '#/definitions/models.Pet'
      petId:
        type: integer
      serviceId:
        type: integer
      services:
        $ref: '#/definitions/models.Service'
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userId:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: vic.pozdeev@gmail.com
    name: Victor Pozdeev
  description: This is API specification for vet-clinic API server.
  license:
    name: MIT
    url: https://opensource.org/licenses/mit-license.php
  title: Vet clinic API
  version: v0.1.0
paths:
...
  /visits:
    get:
      consumes:
      - application/json
      description: Returns the list of visits.
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            items:
              $ref: '#/definitions/models.Visit'
            type: array
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a visit list.
      tags:
      - Visits
    post:
      consumes:
      - application/json
      description: Create a new visit.
      parameters:
      - description: A new visit data for creating.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.VisitDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Visit'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Create a new visit. Required user''s role: Admin'
      tags:
      - Visits
  /visits/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the existing visit.
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Visit'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Delete the existing visit. Required user''s role: Superuser'
      tags:
      - Visits
    get:
      consumes:
      - application/json
      description: Returns one record matched visit's id.
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Visit'
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a visit.
      tags:
      - Visits
    put:
      consumes:
      - application/json
      description: Update the existing visit.
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: string
      - description: Visit data for update.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.VisitDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Visit'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Update the existing visit. Required user''s role: Admin'
      tags:
      - Visits
schemes:
- http
swagger: "2.0"